Move util/featureSelection to selection/
